Remove duplicated FakeApi
