+++ commands/rm_test.go
@@ -6,12 +6,13 @@ import (
 	"github.com/docker/machine/commands/commandstest"
 	"github.com/docker/machine/drivers/fakedriver"
 	"github.com/docker/machine/libmachine/host"
+	"github.com/docker/machine/libmachine/libmachinetest"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestCmdRmMissingMachineName(t *testing.T) {
 	commandLine := &commandstest.FakeCommandLine{}
-	api := &commandstest.FakeLibmachineAPI{}
+	api := &libmachinetest.FakeAPI{}
 
 	err := cmdRm(commandLine, api)
 
@@ -23,7 +24,7 @@ func TestCmdRm(t *testing.T) {
 	commandLine := &commandstest.FakeCommandLine{
 		CliArgs: []string{"machineToRemove1", "machineToRemove2"},
 	}
-	api := &commandstest.FakeLibmachineAPI{
+	api := &libmachinetest.FakeAPI{
 		Hosts: []*host.Host{
 			{
 				Name:   "machineToRemove1",
@@ -43,7 +44,7 @@ func TestCmdRm(t *testing.T) {
 	err := cmdRm(commandLine, api)
 	assert.NoError(t, err)
 
-	assert.False(t, commandstest.Exists(api, "machineToRemove1"))
-	assert.False(t, commandstest.Exists(api, "machineToRemove2"))
-	assert.True(t, commandstest.Exists(api, "machine"))
+	assert.False(t, libmachinetest.Exists(api, "machineToRemove1"))
+	assert.False(t, libmachinetest.Exists(api, "machineToRemove2"))
+	assert.True(t, libmachinetest.Exists(api, "machine"))
 }
